// drr/src/UI/src/model/yesNoOption.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * DRR API
 * OpenAPI spec version: 1.0.0
 */

export type YesNoOption = typeof YesNoOption[keyof typeof YesNoOption];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const YesNoOption = {
  No: 'No',
  Yes: 'Yes',
  NotApplicable: 'NotApplicable',
} as const;